requirements.txt
pyyaml
easyocr
pandas
neo4j
dotenv
opencv-python
fastapi
uvicorn
pydantic
python-multipart